+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/resource/KafkaAgentClient.java
@@ -11,20 +11,18 @@ import io.strimzi.operator.cluster.model.DnsNameGenerator;
 import io.strimzi.operator.common.Reconciliation;
 import io.strimzi.operator.common.ReconciliationLogger;
 import io.strimzi.operator.common.Util;
-import io.strimzi.operator.common.model.PemKeyStoreSupplier;
-import io.strimzi.operator.common.model.PemTrustStoreSupplier;
+import io.strimzi.operator.common.model.PemAuthIdentity;
+import io.strimzi.operator.common.model.PemTrustSet;
 
 import javax.net.ssl.KeyManagerFactory;
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.TrustManagerFactory;
-import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
-import java.nio.charset.StandardCharsets;
 import java.security.GeneralSecurityException;
 import java.security.KeyFactory;
 import java.security.KeyStore;
@@ -33,7 +31,7 @@ import java.security.NoSuchAlgorithmException;
 import java.security.PrivateKey;
 import java.security.cert.Certificate;
 import java.security.cert.CertificateException;
-import java.security.cert.CertificateFactory;
+import java.security.cert.X509Certificate;
 import java.security.spec.InvalidKeySpecException;
 import java.security.spec.PKCS8EncodedKeySpec;
 
@@ -48,21 +46,20 @@ class KafkaAgentClient {
     private static final String BROKER_STATE_REST_PATH = "/v1/broker-state/";
     private static final int BROKER_STATE_HTTPS_PORT = 8443;
     private static final String KEYSTORE_TYPE_JKS = "JKS";
-    private static final String CERT_TYPE_X509 = "X.509";
     private static final char[] KEYSTORE_PASSWORD = "changeit".toCharArray();
     private final String namespace;
     private final Reconciliation reconciliation;
     private final String cluster;
-    private PemTrustStoreSupplier pemTrustStoreSupplier;
-    private PemKeyStoreSupplier pemKeyStoreSupplier;
+    private PemTrustSet pemTrustSet;
+    private PemAuthIdentity pemAuthIdentity;
     private HttpClient httpClient;
 
-    KafkaAgentClient(Reconciliation reconciliation, String cluster, String namespace, PemTrustStoreSupplier pemTrustStoreSupplier, PemKeyStoreSupplier pemKeyStoreSupplier) {
+    KafkaAgentClient(Reconciliation reconciliation, String cluster, String namespace, PemTrustSet pemTrustSet, PemAuthIdentity pemAuthIdentity) {
         this.reconciliation = reconciliation;
         this.cluster = cluster;
         this.namespace = namespace;
-        this.pemTrustStoreSupplier = pemTrustStoreSupplier;
-        this.pemKeyStoreSupplier = pemKeyStoreSupplier;
+        this.pemTrustSet = pemTrustSet;
+        this.pemAuthIdentity = pemAuthIdentity;
         this.httpClient = createHttpClient();
     }
 
@@ -73,7 +70,7 @@ class KafkaAgentClient {
     }
 
     private HttpClient createHttpClient() {
-        if (pemTrustStoreSupplier == null || pemKeyStoreSupplier == null) {
+        if (pemTrustSet == null || pemAuthIdentity == null) {
             throw new RuntimeException("Missing cluster CA and operator certificates required to create connection to Kafka Agent");
         }
 
@@ -98,28 +95,25 @@ class KafkaAgentClient {
     }
 
     private KeyStore getTrustStore() throws CertificateException, KeyStoreException, IOException, NoSuchAlgorithmException {
-        final CertificateFactory caCertFactory = CertificateFactory.getInstance(CERT_TYPE_X509);
-        final Certificate caCert = caCertFactory.generateCertificate(new ByteArrayInputStream(
-                pemTrustStoreSupplier.pemTrustedCertificates().getBytes(StandardCharsets.US_ASCII)));
         KeyStore trustStore = KeyStore.getInstance(KEYSTORE_TYPE_JKS);
         trustStore.load(null);
-        trustStore.setCertificateEntry("ca", caCert);
+        int aliasIndex = 0;
+        for (X509Certificate certificate : pemTrustSet.trustedCertificates()) {
+            trustStore.setEntry(certificate.getSubjectX500Principal().getName() + "-" + aliasIndex, new KeyStore.TrustedCertificateEntry(certificate), null);
+            aliasIndex++;
+        }
         return trustStore;
     }
 
     private KeyStore getKeyStore() throws KeyStoreException, CertificateException, NoSuchAlgorithmException, InvalidKeySpecException, IOException {
-        final CertificateFactory coCertFactory = CertificateFactory.getInstance(CERT_TYPE_X509);
-        final Certificate coCert = coCertFactory.generateCertificate(new ByteArrayInputStream(
-                pemKeyStoreSupplier.pemCertificateChain().getBytes(StandardCharsets.US_ASCII)));
-
-        byte[] decodedKey = Util.decodePemPrivateKey(pemKeyStoreSupplier.pemPrivateKey());
+        byte[] decodedKey = Util.decodePemPrivateKey(pemAuthIdentity.pemPrivateKeyString());
         PKCS8EncodedKeySpec keySpec = new PKCS8EncodedKeySpec(decodedKey);
         final KeyFactory keyFactory = KeyFactory.getInstance("RSA");
         final PrivateKey key = keyFactory.generatePrivate(keySpec);
 
         KeyStore coKeyStore = KeyStore.getInstance(KEYSTORE_TYPE_JKS);
         coKeyStore.load(null);
-        coKeyStore.setKeyEntry("cluster-operator", key, KEYSTORE_PASSWORD, new Certificate[]{coCert});
+        coKeyStore.setKeyEntry("cluster-operator", key, KEYSTORE_PASSWORD, new Certificate[]{pemAuthIdentity.pemCertificateChain()});
 
         return coKeyStore;
     }