+++ operator-common/src/main/java/io/strimzi/operator/common/AdminClientProvider.java
@@ -4,8 +4,8 @@
  */
 package io.strimzi.operator.common;
 
-import io.strimzi.operator.common.model.PemKeyStoreSupplier;
-import io.strimzi.operator.common.model.PemTrustStoreSupplier;
+import io.strimzi.operator.common.model.PemAuthIdentity;
+import io.strimzi.operator.common.model.PemTrustSet;
 import org.apache.kafka.clients.admin.Admin;
 
 import java.util.Properties;
@@ -19,21 +19,21 @@ public interface AdminClientProvider {
      * Create a Kafka Admin interface instance
      *
      * @param bootstrapHostnames Kafka hostname to connect to for administration operations
-     * @param pemTrustStoreSupplier Supplier for fetching the truststore for TLS encryption
-     * @param pemKeyStoreSupplier Supplier for fetching the keystore for TLS client authentication
+     * @param pemTrustSet Trust set for TLS encryption
+     * @param pemAuthIdentity Identity for TLS client authentication
      * @return Instance of Kafka Admin interface
      */
-    Admin createAdminClient(String bootstrapHostnames, PemTrustStoreSupplier pemTrustStoreSupplier, PemKeyStoreSupplier pemKeyStoreSupplier);
+    Admin createAdminClient(String bootstrapHostnames, PemTrustSet pemTrustSet, PemAuthIdentity pemAuthIdentity);
 
     /**
      * Create a Kafka Admin interface instance
      *
      * @param bootstrapHostnames Kafka hostname to connect to for administration operations
-     * @param pemTrustStoreSupplier Supplier for fetching the truststore for TLS encryption
-     * @param pemKeyStoreSupplier Supplier for fetching the keystore for TLS client authentication
+     * @param pemTrustSet Trust set for TLS encryption
+     * @param pemAuthIdentity Identity for TLS client authentication
      * @param config Additional configuration for the Kafka Admin Client
      *
      * @return Instance of Kafka Admin interface
      */
-    Admin createAdminClient(String bootstrapHostnames, PemTrustStoreSupplier pemTrustStoreSupplier, PemKeyStoreSupplier pemKeyStoreSupplier, Properties config);
+    Admin createAdminClient(String bootstrapHostnames, PemTrustSet pemTrustSet, PemAuthIdentity pemAuthIdentity, Properties config);
 }